# pkg/strava/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "The [Swagger Playground](https://developers.strava.com/playground)\
    \ is the easiest way to familiarize yourself with the Strava API by submitting\
    \ HTTP requests and observing the responses before you write any client code.\
    \ It will show what a response will look like with different endpoints depending\
    \ on the authorization scope you receive from your athletes. To use the Playground,\
    \ go to https://www.strava.com/settings/api and change your “Authorization Callback\
    \ Domain” to developers.strava.com. Please note, we only support Swagger 2.0.\
    \ There is a known issue where you can only select one scope at a time. For more\
    \ information, please check the section “client code” at https://developers.strava.com/docs."
  version: "3.0.0"
  title: "Strava API v3"
host: "www.strava.com"
basePath: "/api/v3"
schemes:
- "https"
produces:
- "application/json"
security:
- strava_oauth:
  - "public"
paths:
...
  /athlete/activities:
    get:
      tags:
      - "Activities"
      summary: "List Athlete Activities"
      description: "Returns the activities of an athlete for a specific identifier.\
        \ Requires activity:read. Only Me activities will be filtered out unless requested\
        \ by a token with activity:read_all."
      operationId: "getLoggedInAthleteActivities"
      parameters:
      - name: "before"
        in: "query"
        description: "An epoch timestamp to use for filtering activities that have\
          \ taken place before a certain time."
        required: false
        type: "integer"
        x-exportParamName: "Before"
        x-optionalDataType: "Int32"
      - name: "after"
        in: "query"
        description: "An epoch timestamp to use for filtering activities that have\
          \ taken place after a certain time."
        required: false
        type: "integer"
        x-exportParamName: "After"
        x-optionalDataType: "Int32"
      - name: "page"
        in: "query"
        description: "Page number. Defaults to 1."
        required: false
        type: "integer"
        x-exportParamName: "Page"
        x-optionalDataType: "Int32"
      - name: "per_page"
        in: "query"
        description: "Number of items per page. Defaults to 30."
        required: false
        type: "integer"
        default: 30
        x-exportParamName: "PerPage"
        x-optionalDataType: "Int32"
      responses:
        "200":
          description: "The authenticated athlete's activities"
          examples:
            application/json:
            - resource_state: 2
              athlete:
                id: 134815
                resource_state: 1
              name: "Happy Friday"
              distance: 24931.4
              moving_time: 4500
              elapsed_time: 4500
              total_elevation_gain: 0
              type: "Ride"
              id: 154504250376823
              external_id: "garmin_push_12345678987654321"
              upload_id: 987654321234567891234
              start_date: "2018-05-02T12:15:09Z"
              start_date_local: "2018-05-02T05:15:09Z"
              timezone: "(GMT-08:00) America/Los_Angeles"
              utc_offset: -25200
              location_country: "United States"
              achievement_count: 0
              kudos_count: 3
              comment_count: 1
              athlete_count: 1
              photo_count: 0
              map:
                id: "a12345678987654321"
                resource_state: 2
              trainer: true
              commute: false
              manual: false
              private: false
              flagged: false
              gear_id: "b12345678987654321"
              from_accepted_tag: false
              average_speed: 5.54
              max_speed: 11
              average_cadence: 67.1
              average_watts: 175.3
              weighted_average_watts: 210
              kilojoules: 788.7
              device_watts: true
              has_heartrate: true
              average_heartrate: 140.3
              max_heartrate: 178
              max_watts: 406
              pr_count: 0
              total_photo_count: 1
              has_kudoed: false
              suffer_score: 82
            - resource_state: 2
              athlete:
                id: 167560
                resource_state: 1
              name: "Bondcliff"
              distance: 23676.5
              moving_time: 5400
              elapsed_time: 5400
              total_elevation_gain: 0
              type: "Ride"
              id: 1234567809
              external_id: "garmin_push_12345678987654321"
              upload_id: 1234567819
              start_date: "2018-04-30T12:35:51Z"
              start_date_local: "2018-04-30T05:35:51Z"
              timezone: "(GMT-08:00) America/Los_Angeles"
              utc_offset: -25200
              location_country: "United States"
              achievement_count: 0
              kudos_count: 4
              comment_count: 0
              athlete_count: 1
              photo_count: 0
              map:
                id: "a12345689"
                resource_state: 2
              trainer: true
              commute: false
              manual: false
              private: false
              flagged: false
              gear_id: "b12345678912343"
              from_accepted_tag: false
              average_speed: 4.385
              max_speed: 8.8
              average_cadence: 69.8
              average_watts: 200
              weighted_average_watts: 214
              kilojoules: 1080
              device_watts: true
              has_heartrate: true
              average_heartrate: 152.4
              max_heartrate: 183
              max_watts: 403
              pr_count: 0
              total_photo_count: 1
              has_kudoed: false
              suffer_score: 162
          schema:
            type: "array"
            items:
              $ref: "#/definitions/SummaryActivity"
        default:
          description: "Unexpected error."
          schema:
            $ref: "#/definitions/Fault"
...
securityDefinitions:
  strava_oauth:
    type: "oauth2"
    authorizationUrl: "https://www.strava.com/api/v3/oauth/authorize"
    tokenUrl: "https://www.strava.com/api/v3/oauth/token"
    flow: "accessCode"
    scopes:
      read: "Read public segments, public routes, public profile data, public posts,\
        \ public events, club feeds, and leaderboards"
      read_all: "Read private routes, private segments, and private events for the\
        \ user"
      profile:read_all: "Read all profile information even if the user has set their\
        \ profile visibility to Followers or Only You"
      profile:write: "Update the user's weight and Functional Threshold Power (FTP),\
        \ and access to star or unstar segments on their behalf"
      activity:read: "Read the user's activity data for activities that are visible\
        \ to Everyone and Followers, excluding privacy zone data"
      activity:read_all: "The same access as activity:read, plus privacy zone data\
        \ and access to read the user's activities with visibility set to Only You"
      activity:write: "Access to create manual activities and uploads, and access\
        \ to edit any activities that are visible to the app, based on activity read\
        \ access level"
definitions:
...
  Fault:
    type: "object"
    properties:
      errors:
        type: "array"
        description: "The set of specific errors associated with this fault, if any."
        items:
          $ref: "#/definitions/Error"
      message:
        type: "string"
        description: "The message of the fault."
    description: "Encapsulates the errors that may be returned from the API."
  Error:
    type: "object"
    properties:
      code:
        type: "string"
        description: "The code associated with this error."
      field:
        type: "string"
        description: "The specific field or aspect of the resource associated with\
          \ this error."
      resource:
        type: "string"
        description: "The type of resource associated with this error."
...
  MetaAthlete:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        description: "The unique identifier of the athlete"
...
  ActivityType:
    type: "string"
    description: "An enumeration of the types an activity may have."
    enum:
    - "AlpineSki"
    - "BackcountrySki"
    - "Canoeing"
    - "Crossfit"
    - "EBikeRide"
    - "Elliptical"
    - "Golf"
    - "Handcycle"
    - "Hike"
    - "IceSkate"
    - "InlineSkate"
    - "Kayaking"
    - "Kitesurf"
    - "NordicSki"
    - "Ride"
    - "RockClimbing"
    - "RollerSki"
    - "Rowing"
    - "Run"
    - "Sail"
    - "Skateboard"
    - "Snowboard"
    - "Snowshoe"
    - "Soccer"
    - "StairStepper"
    - "StandUpPaddling"
    - "Surfing"
    - "Swim"
    - "Velomobile"
    - "VirtualRide"
    - "VirtualRun"
    - "Walk"
    - "WeightTraining"
    - "Wheelchair"
    - "Windsurf"
    - "Workout"
    - "Yoga"
...
  LatLng:
    type: "array"
    description: "A pair of latitude/longitude coordinates, represented as an array\
      \ of 2 floating point numbers."
    items:
      type: "number"
      format: "float"
    minItems: 2
    maxItems: 2
...
  PolylineMap:
    type: "object"
    properties:
      id:
        type: "string"
        description: "The identifier of the map"
      polyline:
        type: "string"
        description: "The polyline of the map, only returned on detailed representation\
          \ of an object"
      summary_polyline:
        type: "string"
        description: "The summary polyline of the map"
    example:
      summary_polyline: "summary_polyline"
      id: "id"
      polyline: "polyline"
...
  MetaActivity:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        description: "The unique identifier of the activity"
...
  SummaryActivity:
    allOf:
    - $ref: "#/definitions/MetaActivity"
    - properties:
        external_id:
          type: "string"
          description: "The identifier provided at upload time"
        upload_id:
          type: "integer"
          format: "int64"
          description: "The identifier of the upload that resulted in this activity"
        athlete:
          $ref: "#/definitions/MetaAthlete"
        name:
          type: "string"
          description: "The name of the activity"
        distance:
          type: "number"
          format: "float"
          description: "The activity's distance, in meters"
        moving_time:
          type: "integer"
          description: "The activity's moving time, in seconds"
        elapsed_time:
          type: "integer"
          description: "The activity's elapsed time, in seconds"
        total_elevation_gain:
          type: "number"
          format: "float"
          description: "The activity's total elevation gain."
        elev_high:
          type: "number"
          format: "float"
          description: "The activity's highest elevation, in meters"
        elev_low:
          type: "number"
          format: "float"
          description: "The activity's lowest elevation, in meters"
        type:
          $ref: "#/definitions/ActivityType"
        start_date:
          type: "string"
          format: "date-time"
          description: "The time at which the activity was started."
        start_date_local:
          type: "string"
          format: "date-time"
          description: "The time at which the activity was started in the local timezone."
        timezone:
          type: "string"
          description: "The timezone of the activity"
        start_latlng:
          $ref: "#/definitions/LatLng"
        end_latlng:
          $ref: "#/definitions/LatLng"
        achievement_count:
          type: "integer"
          description: "The number of achievements gained during this activity"
        kudos_count:
          type: "integer"
          description: "The number of kudos given for this activity"
        comment_count:
          type: "integer"
          description: "The number of comments for this activity"
        athlete_count:
          type: "integer"
          description: "The number of athletes for taking part in a group activity"
          minimum: 1
        photo_count:
          type: "integer"
          description: "The number of Instagram photos for this activity"
        total_photo_count:
          type: "integer"
          description: "The number of Instagram and Strava photos for this activity"
        map:
          $ref: "#/definitions/PolylineMap"
        trainer:
          type: "boolean"
          description: "Whether this activity was recorded on a training machine"
        commute:
          type: "boolean"
          description: "Whether this activity is a commute"
        manual:
          type: "boolean"
          description: "Whether this activity was created manually"
        private:
          type: "boolean"
          description: "Whether this activity is private"
        flagged:
          type: "boolean"
          description: "Whether this activity is flagged"
        workout_type:
          type: "integer"
          description: "The activity's workout type"
        upload_id_str:
          type: "string"
          description: "The unique identifier of the upload in string format"
        average_speed:
          type: "number"
          format: "float"
          description: "The activity's average speed, in meters per second"
        max_speed:
          type: "number"
          format: "float"
          description: "The activity's max speed, in meters per second"
        has_kudoed:
          type: "boolean"
          description: "Whether the logged-in athlete has kudoed this activity"
        hide_from_home:
          type: "boolean"
          description: "Whether the activity is muted"
        gear_id:
          type: "string"
          description: "The id of the gear for the activity"
        kilojoules:
          type: "number"
          format: "float"
          description: "The total work done in kilojoules during this activity. Rides\
            \ only"
        average_watts:
          type: "number"
          format: "float"
          description: "Average power output in watts during this activity. Rides\
            \ only"
        device_watts:
          type: "boolean"
          description: "Whether the watts are from a power meter, false if estimated"
        max_watts:
          type: "integer"
          description: "Rides with power meter data only"
        weighted_average_watts:
          type: "integer"
          description: "Similar to Normalized Power. Rides with power meter data only"
...