# pkg/strava/model_activity_type.go
/*
 * Strava API v3
 *
 * The [Swagger Playground](https://developers.strava.com/playground) is the easiest way to familiarize yourself with the Strava API by submitting HTTP requests and observing the responses before you write any client code. It will show what a response will look like with different endpoints depending on the authorization scope you receive from your athletes. To use the Playground, go to https://www.strava.com/settings/api and change your “Authorization Callback Domain” to developers.strava.com. Please note, we only support Swagger 2.0. There is a known issue where you can only select one scope at a time. For more information, please check the section “client code” at https://developers.strava.com/docs.
 *
 * API version: 3.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger
// ActivityType : An enumeration of the types an activity may have.
type ActivityType string

// List of ActivityType
const (
	ALPINE_SKI ActivityType = "AlpineSki"
	BACKCOUNTRY_SKI ActivityType = "BackcountrySki"
	CANOEING ActivityType = "Canoeing"
	CROSSFIT ActivityType = "Crossfit"
	E_BIKE_RIDE ActivityType = "EBikeRide"
	ELLIPTICAL ActivityType = "Elliptical"
	GOLF ActivityType = "Golf"
	HANDCYCLE ActivityType = "Handcycle"
	HIKE ActivityType = "Hike"
	ICE_SKATE ActivityType = "IceSkate"
	INLINE_SKATE ActivityType = "InlineSkate"
	KAYAKING ActivityType = "Kayaking"
	KITESURF ActivityType = "Kitesurf"
	NORDIC_SKI ActivityType = "NordicSki"
	RIDE ActivityType = "Ride"
	ROCK_CLIMBING ActivityType = "RockClimbing"
	ROLLER_SKI ActivityType = "RollerSki"
	ROWING ActivityType = "Rowing"
	RUN ActivityType = "Run"
	SAIL ActivityType = "Sail"
	SKATEBOARD ActivityType = "Skateboard"
	SNOWBOARD ActivityType = "Snowboard"
	SNOWSHOE ActivityType = "Snowshoe"
	SOCCER ActivityType = "Soccer"
	STAIR_STEPPER ActivityType = "StairStepper"
	STAND_UP_PADDLING ActivityType = "StandUpPaddling"
	SURFING ActivityType = "Surfing"
	SWIM ActivityType = "Swim"
	VELOMOBILE ActivityType = "Velomobile"
	VIRTUAL_RIDE ActivityType = "VirtualRide"
	VIRTUAL_RUN ActivityType = "VirtualRun"
	WALK ActivityType = "Walk"
	WEIGHT_TRAINING ActivityType = "WeightTraining"
	WHEELCHAIR ActivityType = "Wheelchair"
	WINDSURF ActivityType = "Windsurf"
	WORKOUT ActivityType = "Workout"
	YOGA ActivityType = "Yoga"
)